pyproject.toml
[tool.poetry]
name = "lasma-birthday-reminders"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^13.13"
Flask = "^2.2.2"
pytz = "^2022.2.1"
cryptography = "^37.0.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
